refactor: change ForWhom to constants
- move forWhom interface to constants
