+++ components/02-molecules/TokensSwapList.tsx
@@ -1,11 +1,11 @@
 import { TokenSwapCard } from "./TokenSwapCard";
-import { ForWhom } from "../03-organisms";
 import {
   TokenCardActionType,
   TokenCardStyleType,
 } from "@/components/02-molecules";
 import { ERC20, EthereumAddress, Token } from "@/lib/shared/types";
 import { EMPTY_ERC_20_BALANCE } from "@/lib/client/blockchain-utils";
+import { ForWhom } from "@/lib/client/constants";
 
 export interface TokensListProps {
   tokensList: Token[];
